starting dark theme

+++ src/services/auth/auth.service.tsx
@@ -1,5 +1,12 @@
 import axios from "axios";
-import { EmailConfirmationResponse, InputsLogin, InputsRegistraytion } from "../../App.types";
+import {
+    EmailConfirmationResponse,
+    InputsChangeData,
+    InputsLogin,
+    InputsRegistraytion,
+    InputsResetPassword
+} from "../../App.types";
+import secureApi from "../../utils/axios-inteseptor/axios-interseptes.ts";
 
 axios.defaults.baseURL = 'https://back.4rooms.pro/api/';
 
@@ -24,9 +31,18 @@ const resetPassword = async (url: string, data: unknown) => {
     });
 };
 
-const logout = () => {
-    localStorage.clear();
-};
+async function logout() {
+    try {
+        const response = await axios.put("logout/");
+        if (response.status === 204) {
+            localStorage.clear();
+            return { success: true };
+        }
+        return { success: false, message: 'Failed to logout' };
+    } catch (error: any) {
+        return { success: false, message: error.response?.data?.detail || 'An error occurred during logout.' };
+    }
+}
 
 async function confirmEmail(token: string): Promise<EmailConfirmationResponse> {
         const response = await axios.get<EmailConfirmationResponse>('https://back.4rooms.pro/api/confirm-email/', {
@@ -40,13 +56,32 @@ const sendSecondEmail = async (email: string) => {
     return await axios.post("send-confirmation-email/", {email});
 };
 
+const changePassword = async (dataForm: InputsResetPassword) => {
+    return secureApi.put('user/change-password/', dataForm)
+}
+const updateUserData = async (data: InputsChangeData) => {
+    return secureApi.put('user/', data);
+};
+
+const updateUserAvatar = async (file:  File | null) => {
+    if (!file) {
+        throw new Error("No file provided");
+    }
+    const formData = new FormData();
+    formData.append('avatar', file);
+    return secureApi.put('profile/', formData);
+};
+
 const authService = {
     signup,
     login,
     resetPassword,
     logout,
     confirmEmail,
-    sendSecondEmail
+    sendSecondEmail,
+    changePassword,
+    updateUserData,
+    updateUserAvatar
 };
 
 export default authService;
